Update publish.yml
It seems that the CI failed due to an older version of Node it's running:

```
error init-package-json@6.0.3: The engine "node" is incompatible with this module. Expected version "^16.14.0 || >=18.0.0". Got "12.22.12"
error Found incompatible module.
```


This change updates node from v12.x to v16.x, the lowest version that this particular package is expecting. We could also bump it up higher (my local environment is at v22.x for example), but whatever you feel comfortable!